Fix failing ITs due to recent features (#4813)
Changes introduced in #4800 to validate compaction configurations caused several ITs to start failing. In these cases the ITs were using invalid configurations. The changes here fix the ITs by either isolating the invalid configurations to only the associated tests (vs having the invalid configuration persist for all the tests in the class), or I made modifications to the test so that they would continue to work.