Stops waiting on scans during tablet close.
Before this change when a tablet was closing it would prevent new scans from starting
and wait for running scans to finish.  This could cause the tablet to become unavailable
for long periods of time.  After this change tablets will no longer wait on running
scans to complete.  They will set something to interrupt the thread and then disable the
client scan session which prevents the client form ever seeing any data after the tablet
is closed.  Once all scan sessions are disabled the tablet will proceed to close.  Disabling
the scan session will cause the client side scanner to eventually switch to the new tablet
server where the tablet is loaded.

Its possible that threads may be left running on the tablet server therefore it will be
important to also implement #4756

fixes #4757