adds startedAfter method to Timer
This [comment][1] identified a flaky unit test. The test was likely flakey because
it compared elapsed times on two timers.  Each call to elapsed time would
call System.nanoTime. Depending on how much nano time drifted between the two
calls different results could be obtained making the code non-deteministic.
Added a method to check if one timer started after another and used it
for this check making the code deterministic.  Its also much easier to understand
the intent of the code with this change.

[1]:https://github.com/apache/accumulo/pull/4821#issuecomment-2307244079