refactors how scan sessions track scan task (#4829)
This is a manual refactor of how scan sessions track scan task.
MultiScanSession and SinceScanSession both extend ScanSession.  Each was
tracking  scan task separately.  Moved this tracking into the parent
class ScanSession.  This refactoring is needed for #4756 but want to
do it on its own to simplify review for bug fix release changes.


Co-authored-by: Christopher L. Shannon <cshannon@apache.org>