Reuse ServiceLock instances
The ServiceLock calls ZooKeeper.exists(path, this) in the constructor,
where "this" is a Watcher. Looking at the places where ServiceLock is
constructed I found several places where this was being done inside of
a loop which may end up creating multiple watchers in the ZooKeeper
client instance. This change moves the ServiceLock construction outside
of the loop.