cleanup scan session when batch scanner closed
When not all data is read from a scanner it can leave server side scan
sessions open.  When the scanner object is closed on the client side
it should close these server side scan sessions.  The batch scanner was
not doing this.  Updated the batch scanner to close sever side sessions.
Added test for the scanner and batch scanner to ensure sessions are
cleaned up.  Improved an existing test related to session timeout.