fixes CloseScannerIT to work with changes in #4840 (#4846)
After the changes in #4840 a scan session with an active thread would
not be discarded.  The change made CloseScannerIT start failing in 3.1.
Adjusted the test to account for this by giving time for deferred
session cleanup that happens when there is an active thread associated
with a scan session.

The test was not failing in 2.1 because the test was less strict in this
branch.  Applied this fix starting in 2.1 to make the test consistent
across branches.