lowers memory burden of listing external compactions in monitor
Each request to the monitor for the list of external compactions would
create a copy of the per compaction information in the
o.a.a.m.r.c.RunningCompaction constructor.  For many concurrent request when
there are lots of external compactions running this could cause memory
problems on the monitor.

This commit changes the code to only create a single RunningCompaction
object every 30 seconds that is used by all request.  This should lower
the amount of memory used as there are concurrent request or even
refreshing the page really frequently.