remove mysterious pass in MetadataConstraints class (#4858)
The Metadata constraint was allowing rows that started with `!!~` a free
pass. AFAIK no code in Accumulo should produce a row with this prefix,
so removed the code.  I suspect this may be somehow related to previous
versions of accumulo where the root table was instead a the root tablet
which was  special tablet in the metadata table.