Avoids listing the sorted logs dir multiple times during log recovery.
The log recovery code would list the sorted walog files multiple times
during recovery.  These changes modify the code to only list the files
once.  Also the listing is cached for a short period of time to improve
the case of multiple tablet referencing the same walogs.  This along
with #4873 should result in much less traffic to the namenode when an
entire accumulo cluster shutsdown and needs to recover.