Update ThriftScanner to use nanotime (#4866)
Update ThriftScanner to use nano time to be consistent with elapsed time
tracking in other locations. Also, convert timeOut to be a Duration to
prevent future errors. This value was being converted to seconds from
millis which is error prone as there was not a good indication the unit
was in seconds.

This is a follow on to #4864