removes bare usage of Text.getBytes()
Calls to Text.getBytes() should always be accompanied with a call to
getLength() because an array that is longer than the data could be
returned if set() was ever called on the Text object.

Replaced with a call to getRowData().toArray() that is more efficient
and does not have the length problem.

Did a partial review of the code and found this one case.  Could be
other cases like this lurking.