switches volumes prior to recovery check (#4889)
In #4873 a check was added to inspect walogs during tablet load to see
if they had any data for the tablet.  This check happens prior to
volume replacement that also runs during tablet load. Therefore if
volume replacement is needed for the walogs then this check will fail
because it can not find the files and the tablet will fail to load.

To fix this problem modified the new check to switch volumes if needed
prior to running the check.


Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>