reduce scan thread contention on native maps (#4910)
Observed many scan threads in the native map code attempting to create
a native map iterator and blocking in registering a cleaner for the
iterator.  The block was happening in jdk.internal.rf.PhantomCleanable.insert()
which has a synchronized code block.  There is a single static cleaner
object on the accumulo code so all scan threads that internact with a
native map could end up contending on this single lock in the process.

This change creates 8 cleaners and therefore 8 locks to reduce the
contention between scan threads.


Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>