Modifies pom to allow git sha in version
Modifies the pom to allow specifying a git sha in the version.  When
testing with downstream software and using a SNAPSHOT version of
accumulo it can be difficult to control what SNAPSHOT you actually get.
These changes allow including the git sha in the accumulo version
number which removes the suprises around using SNAPSHOT versions.

With these changes can include the sha in the version with the following
command.

```
mvn clean install -PskipQA -Dsha1="-$(git rev-parse --short HEAD)"
```

For more information see :

https://maven.apache.org/maven-ci-friendly.html