avoids uneeded object allocation on ServiceLocksPath (#4944)
ServiceLockPaths allows filtering hosts using a predicate.  In the case
where a predicate was passed in that always returned true a HostPort
object needlessly allocated to pass to the predicate.  These changes
refactor the predicate to only allocate the HostAndPort object when
needed.  Before the changes in #4943 this was the behavior, that when
all host were wanted no objects were allocated.