limit thread and add queue to monitor append (#4879)
Add options to configure the AccumuloMonitorAppender as async (default)
with a configurable number of concurrent maxThreads and queueSize. When
the queue is full the monitor appender will now drop log messages.

This implementation uses an executor and thread pool exclusive to the
appender, rather than shared using a global one, which is the default
behavior for HttpClient. The maxThreads and queueSize options only have
an effect if async is true (which is the default).

A user may wish to turn async off for this appender, and wrap this
appender with AsyncAppender instead, which may have additional options
available.

* Additional improvements to monitor appender

* Bump to log4j 2.24.0
* Use the annotation processor to generate Log4j2Plugins.dat in the
  monitor jar's META-INF/, rather than rely on Log4j2's deprecated
  package scanning to find and register the AccumuloMonitorAppender
* Simplify the generics for the AccumuloMonitorAppender.Builder
* Add cleanup to executor, to make a best effort to not leave it running
  when the appender is reconstructed due to configuration changes
* Simplify "canAppend" logic by just saving the executor as a member, so
  we can get the queue size from that
* Logs stats about dropped messages, errors, and total messages sent

Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>