Changes to public API to expose resource groups (#4851)
Added methods to InstanceOperations that return server information (host,
port, resource group, and type) using a new ServerId type. Deprecated older
methods on InstanceOperations that just returned the address of the servers
as Strings. Modified existing code to use the new methods and added a new
test class, InstanceOperationsIT. This work leverages the new ServiceLockPaths
class that was added in #4861 to get information about the servers from their
respective paths in ZooKeeper.

Closes #4849