Set ACCUMULO_SERVICE_INSTANCE consistently for all processes (#4966)
Consistently setting the env variable to include the group
and process number for all processes enables us to remove
code in the other scripts that try and find / apply the
group to various things. The specific issue this solves
is that the PID file name ended up being incorrect when
trying to stop the manager and garbage collector. The
PID file could not be found, because the calculated name
was incorrect, and the processes were not stopped.

Closes #4951