Added TabletServer in WaitForBalanceIT
WaitForBalanceIT was assuming that the tablets were not balanced
after adding splits. In the last step of the split code, in DeleteOperationIds,
the EventCoordinator is notified to try and get the new splits hosted as fast
as possible. This seemed to have created a race condition in the test. Instead,
this commit adds a new TabletServer which will cause a balancing to occur. I
also modified the isBalanced method to take into account the tservers in the
cluster, not just the location from the metadata.