Fixes json extraction in FateOpsCommandsIT
FateOpsCommandsIT ran an external command that would output log messages
and json.  The test wanted to extract only the json from the command
output.  The extraction was broken by a recent change to log messages
that included a curly brace.  Modified the extraction to tolerate this.

These are the same fixes made in #5002, just in a different class.