Modifies ColumnVisibility.normalize() to have no side effects
ColumnVisibility objects have an assoiciated parse tree. Normalizing a
column visibility would change the parse tree in place.  This would
cause the parse tree to no longer match the expression, which would
not matter for most operations on the parse tree.  This side effect
would also cause problems for multiple threads attempting to normalize
a column visibility.  This change creates a new parse tree during the
normalization process and does not modify the existing one.