avoid reading from zookeeper when clearing zoocache entries (#5008)
This change indirectly fixes ZombieScanIT.  That test was creating
scanners running in thread and interrupting those threads.  The
interrupts were routinely being ingnored by code in zoocache that
ignores interrupts when reading from zookeeper.  Looking at the code
that was reading from zookeeper, that code was attempting clear entries
in the cache.  Modified this code to avoid reading from zookeeper when
clearing the cache.