closes reader deep copies outside of lock in FileManager
FileManager has a goal of doing all I/O outside of locks in its code to
avoid one scan blocking another w/ I/O related to opening and closing
files.  The code violated this goal when closing deep copies on an
rfile and did this inside a lock.  This change moves the close outside
of the lock.