Switch getCompactionJob RPC to be async using Jetty
This change switches getCompactionJob() rpc call to be an async REST
call using Jetty and Jersey so that we can use long polling and not
block while waiting for jobs on the queue.

Jetty is configured to use Jackson and a custom serializer/deserializer
to handle Thrift objects being serialized to json.

Authentication has been set up to work with username/password, SSL, and
also Kerberos. Right now for the prototype the authentication is not
quite complete and a couple shortcuts were taken to get it working with
the IT set up so some more work needs to be done before it is finished.