Fix flaky CompactionJobQueuesTest
Fix testGetAsync() by ensuring that future6 will timeout before adding a
job to the queue by waiting for the TimeoutException to happen and not
arbitrarily sleeping. Before it was possible for the job to be added to
the queue in rare cases before the timeout executed so future6 was
receiving the job and not future7 which caused the test to fail.

This closes #5005