adds optional per table metrics
For a subset of metrics in the tablet server and scan server adds
optional tableId tags to meters.  In a follow on change the compactor
could be updated to emit per table metrics, however its current code is
very process oriented and this change should be in its own commit.

Each server process will automatically remove meters for tables that
were delete or related to tables it has not been servicing in a while.