Fixes discrepancy between `admin fate delete` and `UserFateStore.delete` (#5016)
* Fixes discrepancy between `admin fate delete` and `UserFateStore.delete`

`admin fate delete` allows users to delete a transaction. If this transaction is a USER transaction, `UserFateStore.delete()` is how this is achieved. The admin code allows transactions with a status of SUBMITTED, IN_PROGRESS, NEW, FAILED, FAILED_IN_PROGRESS, SUCCESSFUL to call `delete()`, but the `delete()` code only allows NEW, SUBMITTED, SUCCESSFUL, FAILED. Admin now calls a force delete to delete the transaction regardless of state. forceDelete() is a new method only to be used by Admin.