Fix ExternalCompaction_1_IT.testPartialCompaction failure (#5054)
The test was timing out because it never started running compactions
created by the test method. Instead, the compactor process was
running compactions created by the previous test method because
the previous test created a table with a lot of files, started a
user compaction, then cancelled the user compaction. The recent
changes in #5026 caused a bunch of system compactions to be
generated for the table. The two test methods share the same
compaction queue, so the compactor was busy running the system
compactions.

To fix this issue I backported a property added in #3955 that
makes the compactor cancel check method time configurable and
I deleted the table in the test method that created a lot of files.

Closes #5052