Check for tserver failure in balance loop after some iterations
The StatusThread.balanceTablets method could loop indefinitely
in some cases, like when a TabletServer dies but is the target
of a current migration. The status of the TabletServers is
checked before this method is invoked and never updated. Created
a new property to indicate how many iterations should pass before
the Tablet Server status is re-checked.