Reduce ssh connections to hosts in accumulo-cluster (#5066)
The accumulo-cluster script would ssh to the host for each
server that it wanted to start or stop. This change moves
the processing of the number of servers that need to be started
or stopped to the accumulo-service script so that only one
ssh connection needs to be created to the host for that
server type and group.