avoids uneeded work when splitting tablets (#5071)
When tablets needed to split they were queued in a thread pool to seed
a fate operation.  When the tablet group watcher ran again if the tablet
was still in queue it would add it again. This would not cause any
correctness issues, but it would cause a lot of wasted work. Also if one
tablet was queued multiple times it could prevent another tablet that
was not queued at all from being added.  This made splitting very when a
large number of tablets needed to split very inefficient.