Cleanup InstanceOperations API for scans/compactions
This simplifies the API for getting active scans and active compactions.
The new API has been updated to just use a collection of servers instead
of trying to also support other arguments (like a single server). Other
methods were removed if new or deprecated if old, such as
getActiveCompactions().

Getting active scans was updated to be multi-threaded like getting
active compactions, and they now share the same code. If there are not
multiple servers to query then the code now just uses a direct executor
to prevent creating a thread pool.

The listscans command was updated to take advantage of the new API and
to use multiple threads to collect the scans to speed things up.

Lastly, the tests have been updated to account for the API changes.

This closes #5096