Uses more concurrent data structs in new monitor
In the new monitor code SystemInformation class there are
ConcurrentHashMaps that have non concurrent data structs for values.
This was causing problems when multiple threads accessed.  Changed the
values to be concurrent.  Also changed a few other map to be concurrent
just to remove use the normal HashMap import in the class.