Fix flaky FateIT.testTransactionStatus() (#5121)
Related to #2474 and #2550. The sync fix in #2550 helped ensure that the
ZooKeeper client had the updated status, it is still possible for the
node to be deleted before the SUCCESSFUL transaction status is observed.
This change makes FateIT more stable by accepting that as a valid
possible outcome, and no longer fails when that happens.