Enable check of Fate queue size for recommending property change
Added two new properties that control the checking of the Fate
transaction queue size in the background thread that resizes the
Fate thread pool. This new logic will print a warning in the log
that the Fate thread pool size property should be increased if
the number of transactions in the queue is consistently larger
than some multiple of the Fate transaction thread pool size.