Caches non-existence of ZK node for ZooCache.getChildren
For a nodes data in zookeeper ZooCache was caching when a node did not
exist.  However for getChildren when a node did not exist it would not
cache this and would always go to zookeeper. Also ZooCache was handling
a null return from Zookeeper.getChildren that would never happen.  Fixed
these issues.

fixes #5047