Fixed failures in RegexGroupBalanceIT after merging PR #5070 (#5140)
After merging #5070 RegexGroupBalanceIT started failing. Both
GroupBalancer and HostRegexTableLoadBalancer have logic that
throttles the frequency that they can be called do not return
any migrations in this scenario. The change in #5070 modified
the frequency in which the balancer is called from once for all
DataLevel's to once per DataLevel. This caused the GroupBalancer
and HostRegexTableLoadBalancer to return migrations for the ROOT
DataLevel, but not the METADATA and USER DataLevels.

The solution in this commit is to push the DataLevel down to
the Balancer in the BalancerParams so that the throttling can
be done at the DataLevel level.