Validate server types before rpc requests
This is a follow on to #5102

Before making requests such as getScans() or getCompactions() the server
types for each server will be validated to make sure it is a valid type.
The requests are multi-threaded so this pre-check validation prevents
making a bunch of unnecessary RPC server calls if one of the types would
fail validation. This change also fixes the getScans() and
getCompactions() API so that when a server is invalid the correct
exception is thrown and not wrapped in an AccumuloException.