Caches non-existence of ZK node for ZooCache.getChildren (#5143)
* Caches non-existence of ZK node for ZooCache.getChildren

For a nodes data in zookeeper ZooCache was caching when a node did not
exist.  However for getChildren when a node did not exist it would not
cache this and would always go to zookeeper. Also ZooCache was handling
a null return from Zookeeper.getChildren that would never happen.  Fixed
these issues.

fixes #5047

* expire cache entries and remove their watches

* Update test/src/main/java/org/apache/accumulo/test/zookeeper/ZooCacheIT.java

Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>

* code review update

* added test for watch removal and fixed bug with watch removal

* Removes logging added for testing

---------

Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>