Narrows the set of files checked during compaction reservervation
Using the new method added in #5153 the set of files checked in the
conditional mutation for compacaction reserervation was reduced from all
tablet files to only the files in the compaction job.  Noticed the
conditinal mutation was not checking the user compaction requested field
while making this change and added it.