avoids another contentious check in compaction reservation
When reserving a set of files for compaction the conditional mutation
would require the current set of compactions it read in a tablet to be
the same when adding a compaction.  On a busy tablet that has lots of
compactions starting and stopping this set is rapidly changing and this
would cause the conditional mutation to fail and retry.

To avoid this the conditional mutation was improved to check that the
set of files being added for compaction to the tablet do not exist in
any current compactions.  This check will no longer fail for changes to
compactions on unrelated files.