avoid checking all tablet file in bulk import conditional mutation (#5159)
After the changes in #5104 bulk import would read a tablets files,
examine the files, and then require the set of files to be the same to
update the tablet.  The code only cared that the count of files was
below a specified limit. Requiring the set of files not to change was
causing the conditional mutation to fail on a busy tablet.  This change
modifies the condition to check the count of file which avoids
contention on a busy tablet where the set of files is constantly
changing.