narrows check of loaded files in conditional mutation (#5166)
Bulk load fate code was reading a tablets loaded flags, checking it was
not in them, and then making a conditional mutation that required the
set of bulk flags to to be the same.  Requiring the set to be the same
caused unnecessary  collisions between bulk imports.  Modified the
conditional check to require the loaded flags for the fate operation
to be absent.