Add initial TabletMergeability support to client api
This is a follow on PR to #5187 for Issue #5014

This commit adds the first part of the needed client API changes to
support setting TabletMergeability on splits. Two inital updates have
been made to support table creation and adding splits later.

1) A new withSplits() method has been added to NewTableConfiguration
that takes a map of splits and associated TabletMergeability. The
exiting withSplits() method that takes a set just delegates to the new
method and sets a default for TabletMergeability to never because by
default user created tables should never be automatically mergeable.

2) A new putSplits() method has been added that also takes a map and
will create new splits with an associated TabletMergeability. The
existing addSplits() method delgates to this as well. Note that this
method is only partially done. The intent is to support updating
existing splits as well but this will be done in another PR.

The support the changes, where necessary the splits are now encoded with
the TabletMergeability value as json instead of just passing the split
bytes.