Fixes bug with FateInterleavingIT.testInterleaving
Rewrote the test. Seemed to expect an interleave to occur at every opportunity when this may not always occur. Now, just expect at least one interleave to occur. This still accomplishes the same goal of the test: test a fate thread interleaving work on multiple fate ids and ensure the order of operations for any given fate id is as expected even with interleaving.