Base64 encode compaction input file list
The input file list is serialized as json in the compaction
metadata object. When the compaction metadata object was being
serialized into the metadata is too was being serialized as json.
This had the effect of escaping all of the double quotes in the
input file json object because it was json in a String inside of
another json object.

Closes #5060